Separate get_external_db_map
